+++ browser/views/fileManager/styles/section.js
@@ -7,9 +7,9 @@ module.exports = {
   header: css`
     :host {
       background-color: white;
-      font-family: ${fonts.boldSpecial};
-      font-size: 14px;
-
+      font-family: ${fonts.black};
+      font-size: 16px;
+      height: 30px;
       position: sticky;
       top: 0px;
       z-index: 99;