+++ kernel/lib/actionCreators/publish.js
@@ -13,14 +13,14 @@ const {
 const k = require('../../../lib/constants/stateManagement')
 const fs = require('fs')
 const windowManager = require('electron-window-manager')
+const { internalEmitter } = require('electron-window-manager')
 const store = windowManager.sharedData.fetch('store')
 
 ipcMain.on(k.PUBLISH, async (event, load) => {
   debug('%s heard', k.PUBLISH)
   const { password } = store.account
   try {
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: true } })
-    windowManager.pingView({ view: 'filemanager', event: k.REFRESH })
+    internalEmitter.emit(k.CHANGE_PENDING_TRANSACTION_STATE, true)
 
     let dispatchLoad = { name: load.name }
     dispatch({ type: k.FEED_MODAL, load: dispatchLoad })
@@ -28,11 +28,12 @@ ipcMain.on(k.PUBLISH, async (event, load) => {
 
     let { afs: newAFS, afs: { did } } = await afs.create({ owner: load.userAid, password })
     await newAFS.close();
-    (await afs.add({ did, paths: load.paths, password })).close()
+    await (await afs.add({ did, paths: load.paths, password })).close()
 
     const size = load.paths.reduce((sum, file) => sum += fs.statSync(file).size, 0)
     await actionsUtil.writeFileMetaData({ did, size, title: load.name, password })
 
+    debug('Deploying proxy')
     await afs.deploy({ password, did })
 
     debug('Estimating gas')
@@ -55,10 +56,9 @@ ipcMain.on(k.PUBLISH, async (event, load) => {
     dispatch({ type: k.FEED_MODAL, load: dispatchLoad })
     windowManager.closeModal('generalPleaseWaitModal')
     windowManager.openModal('publishConfirmModal')
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: false } })
   } catch (err) {
     debug('Error publishing file %o:', err)
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: false } })
+    internalEmitter.emit(k.CHANGE_PENDING_TRANSACTION_STATE, false)
     windowManager.closeModal('generalPleaseWaitModal')
     dispatch({ type: k.FEED_MODAL, load: { modalName: 'failureModal2' } })
     windowManager.openModal('generalMessageModal')
@@ -70,7 +70,7 @@ ipcMain.on(k.CONFIRM_PUBLISH, async (event, load) => {
   debug('%s heard', k.CONFIRM_PUBLISH)
   const { account, farmer } = store
   try {
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: true } })
+    internalEmitter.emit(k.CHANGE_PENDING_TRANSACTION_STATE, true)
     acmManager.savePublishedItem(load.did, account.userAid)
 
     const descriptorOpts = {
@@ -92,8 +92,7 @@ ipcMain.on(k.CONFIRM_PUBLISH, async (event, load) => {
     const balance = await araContractsManager.getAraBalance(account.userAid)
     debug('Dispatching %s', k.PUBLISHED)
     dispatch({ type: k.PUBLISHED, load: { balance, did: load.did } })
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: false } })
-    windowManager.pingView({ view: 'filemanager', event: k.REFRESH })
+    internalEmitter.emit(k.CHANGE_PENDING_TRANSACTION_STATE, false)
 
     const publishedSub = await araContractsManager.subscribePublished({ did: load.did })
     const rewardsSub = await araContractsManager.subscribeRewardsAllocated(load.did, account.accountAddress, account.userAid, )
@@ -115,8 +114,7 @@ ipcMain.on(k.CONFIRM_PUBLISH, async (event, load) => {
     acmManager.removedPublishedItem(load.did, account.userAid)
     dispatch({ type: k.ERROR_PUBLISHING })
 
-    dispatch({ type: k.CHANGE_PENDING_TRANSACTION_STATE, load: { pendingTransaction: false } })
-    windowManager.pingView({ view: 'filemanager', event: k.REFRESH })
+    internalEmitter.emit(k.CHANGE_PENDING_TRANSACTION_STATE, false)
     windowManager.closeModal('generalPleaseWaitModal')
     //Needs short delay. Race conditions cause modal state to dump after its loaded
     setTimeout(() => {